+++ lib/super/pagination.rb
@@ -82,46 +82,4 @@ module Super
       end
     end
   end
-
-  class Controls
-    module Optional
-      # Specifies how many records to show per page
-      #
-      # @param action [ActionInquirer]
-      # @param query_params [Hash]
-      # @return [ActiveRecord::Relation]
-      def records_per_page(action:, query_params:)
-        Super.configuration.index_records_per_page
-      end
-    end
-
-    module Steps
-      # Sets up pagination
-      #
-      # @param action [ActionInquirer]
-      # @param records [ActiveRecord::Relation]
-      # @param query_params [Hash]
-      # @return [Pagination]
-      def initialize_pagination(action:, records:, query_params:)
-        Pagination.new(
-          total_count: records.size,
-          limit: records_per_page(action: action, query_params: query_params),
-          query_params: query_params,
-          page_query_param: :page
-        )
-      end
-
-      # Paginates
-      #
-      # @param action [ActionInquirer]
-      # @param records [ActiveRecord::Relation]
-      # @param pagination [Pagination]
-      # @return [ActiveRecord::Relation]
-      def paginate_records(action:, records:, pagination:)
-        records
-          .limit(pagination.limit)
-          .offset(pagination.offset)
-      end
-    end
-  end
 end